refactor: complete App Router migration
The codebase carried intermediate client page components
(e.g. HomePageClient, DiscoverPageClient) that existed
solely to bridge Pages Router conventions into the App
Router shell. Removing them lets server components own
data fetching and metadata directly.

Route groups split layouts by visual intent: `(contained)`
for width-constrained pages with sidebar, and the default
group for fullscreen map views. This replaces the single
monolithic Layout component.

Server-side auth token forwarding ensures private maps and
reviews resolve correctly on first render, while
global-not-found.tsx handles unmatched routes without
depending on the layout tree.
