+++ next.config.js
@@ -11,5 +11,8 @@ const withSerwist = withSerwistInit({
  * @type {import('next').NextConfig}
  */
 module.exports = withSerwist({
-  output: 'standalone'
+  output: 'standalone',
+  experimental: {
+    globalNotFound: true
+  }
 });
