+++ src/app/[lang]/(contained)/page.tsx
@@ -1,7 +1,9 @@
 import type { Metadata } from 'next';
-import { getPopularMaps } from '../../../lib/maps';
+import Timeline from '../../../components/home/Timeline';
+import TrendingReviews from '../../../components/home/TrendingReviews';
+import { getServerAuthState } from '../../../lib/auth';
+import { getPopularReviews } from '../../../lib/reviews';
 import { getDictionary } from '../../../utils/getDictionary';
-import TermsPageClient from './TermsPageClient';
 
 type Props = {
   params: Promise<{ lang: string }>;
@@ -10,7 +12,7 @@ type Props = {
 export async function generateMetadata({ params }: Props): Promise<Metadata> {
   const { lang } = await params;
   const dict = getDictionary(lang);
-  const title = `${dict['terms of service']} | Qoodish`;
+  const title = 'Qoodish';
   const description = dict['meta description'];
   const thumbnailUrl =
     lang === 'en'
@@ -24,17 +26,17 @@ export async function generateMetadata({ params }: Props): Promise<Metadata> {
     keywords:
       'Qoodish, qoodish, 食べ物, グルメ, 食事, マップ, 地図, 友だち, グループ, 旅行, 観光, maps, travel, food, group, trip',
     alternates: {
-      canonical: `${endpoint}/${lang}/terms`,
+      canonical: `${endpoint}/${lang}`,
       languages: {
-        en: `${endpoint}/en/terms`,
-        ja: `${endpoint}/ja/terms`,
-        'x-default': `${endpoint}/en/terms`
+        en: `${endpoint}/en`,
+        ja: `${endpoint}/ja`,
+        'x-default': `${endpoint}/en`
       }
     },
     openGraph: {
       title,
       description,
-      url: `${endpoint}/${lang}/terms`,
+      url: `${endpoint}/${lang}`,
       images: [{ url: thumbnailUrl }],
       locale: lang === 'en' ? 'en_US' : 'ja_JP',
       siteName: dict['meta headline']
@@ -45,9 +47,14 @@ export async function generateMetadata({ params }: Props): Promise<Metadata> {
   };
 }
 
-export default async function TermsPage({ params }: Props) {
+export default async function HomePage({ params }: Props) {
   const { lang } = await params;
-  const popularMaps = await getPopularMaps(lang);
+  const { authenticated } = await getServerAuthState();
+  const popularReviews = authenticated ? [] : await getPopularReviews(lang);
 
-  return <TermsPageClient popularMaps={popularMaps} />;
+  return authenticated ? (
+    <Timeline />
+  ) : (
+    <TrendingReviews reviews={popularReviews} />
+  );
 }
