--- src/app/[lang]/settings/page.tsx
@@ -1,53 +1,0 @@
-import type { Metadata } from 'next';
-import { getPopularMaps } from '../../../lib/maps';
-import { getDictionary } from '../../../utils/getDictionary';
-import SettingsPageClient from './SettingsPageClient';
-
-type Props = {
-  params: Promise<{ lang: string }>;
-};
-
-export async function generateMetadata({ params }: Props): Promise<Metadata> {
-  const { lang } = await params;
-  const dict = getDictionary(lang);
-  const title = `${dict.settings} | Qoodish`;
-  const description = dict['meta description'];
-  const thumbnailUrl =
-    lang === 'en'
-      ? process.env.NEXT_PUBLIC_OGP_IMAGE_URL_EN
-      : process.env.NEXT_PUBLIC_OGP_IMAGE_URL_JA;
-  const endpoint = process.env.NEXT_PUBLIC_ENDPOINT;
-
-  return {
-    title,
-    description,
-    keywords:
-      'Qoodish, qoodish, 食べ物, グルメ, 食事, マップ, 地図, 友だち, グループ, 旅行, 観光, maps, travel, food, group, trip',
-    alternates: {
-      canonical: `${endpoint}/${lang}/settings`,
-      languages: {
-        en: `${endpoint}/en/settings`,
-        ja: `${endpoint}/ja/settings`,
-        'x-default': `${endpoint}/en/settings`
-      }
-    },
-    openGraph: {
-      title,
-      description,
-      url: `${endpoint}/${lang}/settings`,
-      images: [{ url: thumbnailUrl }],
-      locale: lang === 'en' ? 'en_US' : 'ja_JP',
-      siteName: dict['meta headline']
-    },
-    twitter: {
-      card: 'summary_large_image'
-    }
-  };
-}
-
-export default async function SettingsPage({ params }: Props) {
-  const { lang } = await params;
-  const popularMaps = await getPopularMaps(lang);
-
-  return <SettingsPageClient popularMaps={popularMaps} />;
-}
