+++ src/components/auth/LoginCard.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import {
   Box,
   Card,
