+++ src/components/home/Timeline.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { Reviews } from '@mui/icons-material';
 import { Box, Button, CircularProgress, Stack } from '@mui/material';
 import { memo, useCallback, useState } from 'react';
