+++ src/components/home/TrendingReviews.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { Reviews } from '@mui/icons-material';
 import { Box, Button, Stack, Typography } from '@mui/material';
 import Link from 'next/link';
