--- src/components/layouts/BackButton.tsx
@@ -1,22 +1,0 @@
-import { ArrowBack } from '@mui/icons-material';
-import { IconButton } from '@mui/material';
-import { useRouter } from 'next/navigation';
-import { memo, useCallback } from 'react';
-
-export default memo(function BackButton() {
-  const { back, push } = useRouter();
-
-  const handleBackButtonClick = useCallback(() => {
-    if (window.history.length > 2) {
-      back();
-    } else {
-      push('/');
-    }
-  }, [back, push]);
-
-  return (
-    <IconButton color="inherit" onClick={handleBackButtonClick} edge="start">
-      <ArrowBack />
-    </IconButton>
-  );
-});
