+++ src/components/layouts/Footer.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import {
   Box,
   Container,
