+++ src/components/layouts/MobileAppBar.tsx
@@ -1,25 +1,18 @@
+'use client';
+
 import { Search } from '@mui/icons-material';
 import { AppBar, Box, IconButton, Toolbar } from '@mui/material';
 import { memo, useContext, useState } from 'react';
 import AuthContext from '../../context/AuthContext';
+import ShellContext from '../../context/ShellContext';
 import { useProfile } from '../../hooks/useProfile';
 import ProfileAvatar from '../common/ProfileAvatar';
-import BackButton from './BackButton';
 import Logo from './Logo';
 import MobileDrawer from './MobileDrawer';
 
-type Props = {
-  showBackButton?: boolean;
-  onSearchOpen: () => void;
-  onCreateMapClick: () => void;
-};
-
-export default memo(function MobileAppBar({
-  showBackButton,
-  onSearchOpen,
-  onCreateMapClick
-}: Props) {
+export default memo(function MobileAppBar() {
   const { currentUser } = useContext(AuthContext);
+  const { openSearch, openCreateMap } = useContext(ShellContext);
   const { profile } = useProfile(currentUser?.uid);
 
   const [drawerOpen, setDrawerOpen] = useState(false);
@@ -34,23 +27,19 @@ export default memo(function MobileAppBar({
           }}
         >
           <Box sx={{ display: 'flex', justifyContent: 'flex-start' }}>
-            {showBackButton ? (
-              <BackButton />
-            ) : (
-              <IconButton
-                size="small"
-                edge="start"
-                onClick={() => setDrawerOpen(true)}
-              >
-                <ProfileAvatar profile={profile} size={32} />
-              </IconButton>
-            )}
+            <IconButton
+              size="small"
+              edge="start"
+              onClick={() => setDrawerOpen(true)}
+            >
+              <ProfileAvatar profile={profile} size={32} />
+            </IconButton>
           </Box>
 
           <Logo color="inherit" />
 
           <Box sx={{ display: 'flex', justifyContent: 'flex-end' }}>
-            <IconButton onClick={onSearchOpen} edge="end" color="inherit">
+            <IconButton onClick={openSearch} edge="end" color="inherit">
               <Search />
             </IconButton>
           </Box>
@@ -61,7 +50,7 @@ export default memo(function MobileAppBar({
         open={drawerOpen}
         onOpen={() => setDrawerOpen(true)}
         onClose={() => setDrawerOpen(false)}
-        onCreateMapClick={onCreateMapClick}
+        onCreateMapClick={openCreateMap}
       />
     </>
   );
