+++ src/components/notifications/NotificationList.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { Notifications } from '@mui/icons-material';
 import {
   Avatar,
