+++ src/components/settings/PushNotificationsCard.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import {
   Button,
   Card,
