+++ src/lib/reviews.ts
@@ -2,20 +2,27 @@ import type { Review } from '../../types';
 
 export async function getReview(
   reviewId: string,
-  lang: string
+  lang: string,
+  token?: string,
+  mapId?: string
 ): Promise<Review | null> {
   try {
-    const res = await fetch(
-      `${process.env.API_ENDPOINT}/guest/reviews/${reviewId}`,
-      {
-        headers: {
-          Accept: 'application/json',
-          'Accept-Language': lang,
-          'Content-Type': 'application/json'
-        },
-        next: { revalidate: 300 }
-      }
-    );
+    const path =
+      token && mapId
+        ? `/maps/${mapId}/reviews/${reviewId}`
+        : `/guest/reviews/${reviewId}`;
+    const headers: Record<string, string> = {
+      Accept: 'application/json',
+      'Accept-Language': lang,
+      'Content-Type': 'application/json'
+    };
+    if (token) {
+      headers.Authorization = `Bearer ${token}`;
+    }
+    const res = await fetch(`${process.env.API_ENDPOINT}${path}`, {
+      headers,
+      next: { revalidate: token ? 0 : 300 }
+    });
     if (!res.ok) {
       console.error(`Failed to fetch review ${reviewId}: ${res.status}`);
       return null;
